edit code for news

+++ protected/controllers/admin/DetailMenuController.php
@@ -26,23 +26,20 @@ class DetailMenuController extends Controller
 	 */
 	public function accessRules()
 	{
-		return array(
-			array('allow',  // allow all users to perform 'index' and 'view' actions
-				'actions'=>array('index','view'),
-				'users'=>array('*'),
-			),
-			array('allow', // allow authenticated user to perform 'create' and 'update' actions
-				'actions'=>array('create','update'),
-				'users'=>array('@'),
-			),
-			array('allow', // allow admin user to perform 'admin' and 'delete' actions
-				'actions'=>array('admin','delete'),
-				'users'=>array('admin'),
-			),
-			array('deny',  // deny all users
-				'users'=>array('*'),
-			),
-		);
+    if(app()->user->getState('roles') =="admin") {
+
+      $arr =array('index','create', 'update', 'view', 'admin', 'delete');   /* give all access to admin */
+    } else {
+
+      $arr = array('view', 'admin', 'index');    /*  no access to other user */
+    }
+
+    return array(array('allow',
+      'actions'=>$arr,
+      'users'=>array('@'),),
+      array('deny',
+        'users'=>array('*'),),
+    );
 	}
   /*
        *  init CSS and Javascript file
@@ -170,7 +167,10 @@ class DetailMenuController extends Controller
         if (is_object($image_path) && get_class($image_path)==='CUploadedFile')
         {
           if(is_object($image_path)) {
-            unlink(Yii::getPathOfAlias('webroot') . detailMenu::S_THUMBNAIL . $old_image_path);
+            if($old_image_path != '') {
+              unlink(Yii::getPathOfAlias('webroot') . detailMenu::S_THUMBNAIL . $old_image_path);
+            }
+
             $model->image_path = $image_path;
           } else {
             $model->image_path = $old_image_path;
